requirements.txt
flask==2.3.3
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
plotly==5.17.0
dash==2.14.1
dash-bootstrap-components==1.5.0
requests==2.31.0
joblib==1.3.2
gunicorn==21.2.0
python-dotenv==1.0.0
openai==0.28.1
folium==0.14.0